LogisticLinearRegression in blackbox umbenannt
